[IMP] mrp: simplify & speed up tests common
